changed to cast to Float to see if the error is fixed

+++ superflyingthing-core/src/main/java/com/gemserk/games/superflyingthing/Scripts.java
@@ -584,7 +584,7 @@ public class Scripts {
 			}
 			
 			SpriteComponent spriteComponent = ComponentWrapper.getSpriteComponent(e);
-			spriteComponent.getColor().a = laserTimelineAnimation.getValue("alpha");
+			spriteComponent.getColor().a = (Float) laserTimelineAnimation.getValue("alpha");
 			
 			Physics physics = ComponentWrapper.getPhysics(e);
 			Contact contact = physics.getContact();
